8.19/henaw_implementation/requirements.txt
# HENAW Model Requirements
# PyTorch implementation for biological age prediction

# Core dependencies
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
pyyaml>=5.4.0

# Data handling
h5py>=3.0.0

# Evaluation and metrics
lifelines>=0.27.0  # For survival analysis

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Interpretability
shap>=0.40.0
captum>=0.5.0  # PyTorch model interpretability

# Logging and monitoring
tensorboard>=2.8.0
wandb>=0.12.0
tqdm>=4.62.0

# Optional: API server
flask>=2.0.0  # For REST API server

# Optional: Performance optimization
onnx>=1.10.0  # For model export
onnxruntime>=1.9.0  # For optimized inference

# Development tools (optional)
pytest>=6.2.0
black>=21.0
flake8>=3.9.0
mypy>=0.900